refacto: read from S3 instead of REST api (#332)
* wip: refacto to read from S3 instead of REST api

* wip: read from parquet s3 done, see if we update programs there or not

* Potential fix for code scanning alert no. 1: Clear-text logging of sensitive information


* wip: read and update programs and filters

* wip: test

* wip: test

* review

---------

+++ postgres/schemas/models.py
@@ -247,10 +247,11 @@ def update_program_metadata(engine):
     except (Exception) as error:
         logging.error(f"Error : Update program metadata {error}")
 
-def empty_tables(session = None):
+def empty_tables(session = None, stop_word = True):
     if(os.environ.get("POSTGRES_HOST") == "postgres_db" or os.environ.get("POSTGRES_HOST") == "localhost"):
         logging.warning("""Doing: Empty table Stop_Word / Keywords""")
-        session.query(Stop_Word).delete()
+        if stop_word:
+            session.query(Stop_Word).delete()
         session.query(Keywords).delete()
         session.commit()
         logging.warning("""Done: Empty table Stop_Word / Keywords""")
